changes
changes

+++ Models/Response/CategoryDetailsResponse.cs
@@ -7,14 +7,14 @@
         public CategoryDetailsParentId Parent_Id { get; set; }
         public int Box { get; set; }
         public DateTime Entry_Date { get; set; }
-        public int Sq_Mt { get; set; }
+        public double Sq_Mt { get; set; }
     }
 
     public class CategoryDetailsDatum
     {
         public List<CategoryDetailsResponse> Category_Details_from_APP { get; set; }
         public string id { get; set; }
-        public int Total_Sq_Mt { get; set; }
+        public double Total_Sq_Mt { get; set; }
     }
 
     public class CategoryDetailsParentId
